# common-platform-distributor-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-distributor-mgmt
  description: Distributor Management Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/distributors
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/distributors/{distributorId}/brandings/{brandingId}:
    get:
      tags:
        - Distributor Branding
      summary: Get distributor branding by ID
      description: Returns a distributor branding based on its ID
      operationId: getDistributorBrandingById
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: integer
            format: int64
        - name: brandingId
          in: path
          description: ID of the distributor branding to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved distributor branding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorBrandingDTO'
        '404':
          description: Distributor branding not found
        '500':
          description: Internal server error
    put:
      tags:
        - Distributor Branding
      summary: Update distributor branding
      description: Updates an existing distributor branding with the provided information
      operationId: updateDistributorBranding
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: integer
            format: int64
        - name: brandingId
          in: path
          description: ID of the distributor branding to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorBrandingDTO'
        required: true
      responses:
        '200':
          description: Distributor branding successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorBrandingDTO'
        '400':
          description: Invalid distributor branding data provided
        '404':
          description: Distributor branding not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Distributor Branding
      summary: Delete distributor branding
      description: Deletes a distributor branding based on its ID
      operationId: deleteDistributorBranding
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: integer
            format: int64
        - name: brandingId
          in: path
          description: ID of the distributor branding to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Distributor branding successfully deleted
        '404':
          description: Distributor branding not found
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}:
    get:
      tags:
        - Distributor
      summary: Get distributor by ID
      description: Returns a distributor based on its ID
      operationId: getDistributorById
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved distributor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorDTO'
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
    put:
      tags:
        - Distributor
      summary: Update distributor
      description: Updates an existing distributor with the provided information
      operationId: updateDistributor
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorDTO'
        required: true
      responses:
        '200':
          description: Distributor successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorDTO'
        '400':
          description: Invalid distributor data provided
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Distributor
      summary: Delete distributor
      description: Deletes a distributor based on its ID
      operationId: deleteDistributor
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Distributor successfully deleted
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}/brandings/filter:
    post:
      tags:
        - Distributor Branding
      summary: Filter distributor brandings
      description: Returns a paginated list of distributor brandings based on filter criteria
      operationId: filterDistributorBrandings
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestDistributorBrandingDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved distributor brandings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria provided
        '500':
          description: Internal server error
  /api/v1/distributors/{distributorId}/brandings:
    post:
      tags:
        - Distributor Branding
      summary: Create a new distributor branding
      description: Creates a new distributor branding with the provided information
      operationId: createDistributorBranding
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorBrandingDTO'
        required: true
      responses:
        '201':
          description: Distributor branding successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorBrandingDTO'
        '400':
          description: Invalid distributor branding data provided
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
...
  /api/v1/distributors/filter:
    post:
      tags:
        - Distributor
      summary: Filter distributors
      description: Returns a paginated list of distributors based on filter criteria
      operationId: filterDistributors
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestDistributorDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved distributors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria provided
        '500':
          description: Internal server error
...
  /api/v1/leasing-contracts/distributor/{distributorId}:
    get:
      tags:
        - Leasing Contract
      summary: Get leasing contracts by distributor
      description: Returns all leasing contracts associated with a specific distributor
      operationId: getLeasingContractsByDistributorId
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved leasing contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeasingContractDTO'
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
...
components:
  schemas:
    JsonNode:
      type: object
...
    LeasingContractDTO:
      required:
        - contractId
        - distributorId
        - downPayment
        - endDate
        - lendingConfigurationId
        - monthlyPayment
        - partyId
        - productId
        - startDate
        - totalAmount
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        contractId:
          type: integer
          format: int64
        partyId:
          type: integer
          format: int64
        distributorId:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        lendingConfigurationId:
          type: integer
          format: int64
        product:
          $ref: '#/components/schemas/ProductDTO'
        lendingConfiguration:
          $ref: '#/components/schemas/LendingConfigurationDTO'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        monthlyPayment:
          type: number
        downPayment:
          type: number
        totalAmount:
          type: number
        status:
          pattern: DRAFT|PENDING|APPROVED|ACTIVE|COMPLETED|CANCELLED|TERMINATED
          type: string
        approvalDate:
          type: string
          format: date-time
        approvedBy:
          type: integer
          format: int64
        termsConditions:
          maxLength: 5000
          minLength: 0
          type: string
        notes:
          maxLength: 1000
          minLength: 0
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: integer
          format: int64
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        updatedBy:
          type: integer
          format: int64
          readOnly: true
    LendingConfigurationDTO:
      required:
        - defaultTermMonths
        - lendingType
        - maxTermMonths
        - minTermMonths
        - name
        - productId
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        productId:
          type: integer
          format: int64
        name:
          maxLength: 255
          minLength: 0
          type: string
        description:
          maxLength: 1000
          minLength: 0
          type: string
        lendingType:
          $ref: '#/components/schemas/LendingTypeDTO'
        minTermMonths:
          type: integer
          format: int32
        maxTermMonths:
          type: integer
          format: int32
        defaultTermMonths:
          type: integer
          format: int32
        minDownPaymentPercentage:
          type: number
        defaultDownPaymentPercentage:
          type: number
        interestRate:
          type: number
        processingFeePercentage:
          type: number
        earlyTerminationFeePercentage:
          type: number
        latePaymentFeeAmount:
          type: number
        gracePeriodDays:
          minimum: 0
          type: integer
          format: int32
        isDefault:
          type: boolean
        isActive:
          type: boolean
        termsConditions:
          maxLength: 5000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
    LendingTypeDTO:
      required:
        - code
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          maxLength: 255
          minLength: 0
          type: string
        code:
          maxLength: 100
          minLength: 0
          type: string
        description:
          maxLength: 1000
          minLength: 0
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: integer
          format: int64
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        updatedBy:
          type: integer
          format: int64
          readOnly: true
    ProductCategoryDTO:
      required:
        - code
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          maxLength: 255
          minLength: 0
          type: string
        code:
          maxLength: 100
          minLength: 0
          type: string
        description:
          maxLength: 1000
          minLength: 0
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: integer
          format: int64
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        updatedBy:
          type: integer
          format: int64
          readOnly: true
    ProductDTO:
      required:
        - distributorId
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        distributorId:
          type: integer
          format: int64
        name:
          maxLength: 255
          minLength: 0
          type: string
        description:
          maxLength: 1000
          minLength: 0
          type: string
        sku:
          maxLength: 100
          minLength: 0
          type: string
        modelNumber:
          maxLength: 100
          minLength: 0
          type: string
        manufacturer:
          maxLength: 255
          minLength: 0
          type: string
        category:
          $ref: '#/components/schemas/ProductCategoryDTO'
        imageUrl:
          maxLength: 255
          minLength: 0
          type: string
        specifications:
          $ref: '#/components/schemas/JsonNode'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
...
    DistributorDTO:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        externalCode:
          maxLength: 255
          minLength: 0
          type: string
        name:
          maxLength: 255
          minLength: 0
          type: string
        displayName:
          maxLength: 255
          minLength: 0
          type: string
        taxId:
          type: string
        registrationNumber:
          maxLength: 100
          minLength: 0
          type: string
        websiteUrl:
          maxLength: 255
          minLength: 0
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        supportEmail:
          type: string
        addressLine:
          maxLength: 255
          minLength: 0
          type: string
        postalCode:
          maxLength: 20
          minLength: 0
          type: string
        city:
          maxLength: 100
          minLength: 0
          type: string
        state:
          maxLength: 100
          minLength: 0
          type: string
        countryCode:
          pattern: ^[A-Z]{2}$
          type: string
        isActive:
          type: boolean
        isTestDistributor:
          type: boolean
        timeZone:
          maxLength: 50
          minLength: 0
          type: string
        defaultLocale:
          maxLength: 10
          minLength: 0
          type: string
        onboardedAt:
          type: string
          format: date-time
        terminatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
...
    DistributorBrandingDTO:
      required:
        - distributorId
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        distributorId:
          type: integer
          format: int64
        logoUrl:
          maxLength: 255
          minLength: 0
          type: string
        faviconUrl:
          maxLength: 255
          minLength: 0
          type: string
        primaryColor:
          pattern: ^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$
          type: string
        secondaryColor:
          pattern: ^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$
          type: string
        backgroundColor:
          pattern: ^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$
          type: string
        fontFamily:
          maxLength: 100
          minLength: 0
          type: string
        theme:
          type: string
          enum:
            - LIGHT
            - DARK
            - CUSTOM
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestDistributorBrandingDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/DistributorBrandingDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...
    FilterRequestDistributorDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/DistributorDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination